ipacm: Fix vlan_id and associated_VIDs variables type issue
Fix the type of vlan_id and associated_VIDs to
uint16_t for larger than 256 value support.

(cherry picked from commit 5a574deaae203251d3a998b9707eadddd153b457)

+++ ipacm/inc/IPACM_Wan.h
@@ -127,7 +127,7 @@ typedef struct
 	uint32_t ipv4_addr;
 	bool wan_up_vlan;
 	bool is_xlat;
-	uint8_t associated_VIDs[IPA_MAX_NUM_SW_PDNS];
+	uint16_t associated_VIDs[IPA_MAX_NUM_SW_PDNS];
 	uint8_t VID_cnt = 0;
 	IPACM_Wan *pIface;
 }ipacm_ipv4_wan_iface;
@@ -136,7 +136,7 @@ typedef struct
 {
 	uint32_t ipv6_prefix[2];
 	bool wan_up_vlan_v6;
-	uint8_t associated_VIDs[IPA_MAX_NUM_SW_PDNS];
+	uint16_t associated_VIDs[IPA_MAX_NUM_SW_PDNS];
 	uint8_t VID_cnt = 0;
 	IPACM_Wan *pIface;
 }ipacm_ipv6_wan_iface;
